requirements.txt
flask-httpauth
boto3
gunicorn
Flask
flask-cors
flask-sqlalchemy
python-dotenv
pandas
scikit-learn
joblib
